extract PipelineStage typings from index.d.ts add tests for $addFields and $bucket
